# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Dragon Ball FCT",
        "description": "Documentación oficial de la API. ¡Siente el Ki!",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://api-dragon-ball-fct.test",
            "description": "Servidor Local"
        }
    ],
    "paths": {
        "/api/personajes": {
            "get": {
                "tags": [
                    "Personajes"
                ],
                "summary": "Listar todos los personajes",
                "operationId": "81b94e43f5695c010db6746c025e8588",
                "responses": {
                    "200": {
                        "description": "Lista de personajes obtenida con éxito"
                    }
                }
            },
            "post": {
                "tags": [
                    "Personajes"
                ],
                "summary": "Crear un nuevo personaje",
                "operationId": "bb37a7570de4b6e1946f15ab0eaad748",
                "requestBody": {
                    "description": "Datos del personaje, incluyendo su planeta y transformaciones",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "ki",
                                    "maxKi",
                                    "race",
                                    "gender",
                                    "description",
                                    "image",
                                    "affiliation",
                                    "originPlanet"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Hit"
                                    },
                                    "ki": {
                                        "type": "string",
                                        "example": "10000000"
                                    },
                                    "maxKi": {
                                        "type": "string",
                                        "example": "50000000"
                                    },
                                    "race": {
                                        "type": "string",
                                        "example": "Assassin"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "example": "Male"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Asesino legendario del Universo 6"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "https://ejemplo.com/hit.webp"
                                    },
                                    "affiliation": {
                                        "type": "string",
                                        "example": "Universe 6"
                                    },
                                    "originPlanet": {
                                        "required": [
                                            "name",
                                            "isDestroyed",
                                            "description",
                                            "image"
                                        ],
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Planeta Sadala"
                                            },
                                            "isDestroyed": {
                                                "type": "boolean",
                                                "example": false
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Hogar original de los Saiyans"
                                            },
                                            "image": {
                                                "type": "string",
                                                "example": "https://ejemplo.com/sadala.webp"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "transformations": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Orange Piccolo"
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "https://ejemplo.com/super-saiyan.webp"
                                                },
                                                "ki": {
                                                    "type": "string",
                                                    "example": "99999999"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Personaje creado con éxito"
                    },
                    "422": {
                        "description": "Error de validación (ej. El personaje ya existe o el planeta no es válido)"
                    }
                }
            }
        },
        "/api/personajes/{id}": {
            "get": {
                "tags": [
                    "Personajes"
                ],
                "summary": "Obtener un personaje por su ID",
                "operationId": "df74aada7e99eb3739537ddd8ec04819",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "El ID numérico del personaje que quieres buscar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Personaje encontrado y devuelto con éxito"
                    },
                    "404": {
                        "description": "Personaje no encontrado (o eliminado)"
                    }
                }
            },
            "put": {
                "tags": [
                    "Personajes"
                ],
                "summary": "Actualizar un personaje existente",
                "operationId": "3e32c2eb8b0bc16d87fbf8eb23ef41ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "El ID numérico del personaje que vas a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Envía solo los campos que quieras actualizar. (Ejemplo: Solo enviar el Ki y la descripción)",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Goku (Post-Entrenamiento)"
                                    },
                                    "ki": {
                                        "type": "string",
                                        "example": "90000000"
                                    },
                                    "maxKi": {
                                        "type": "string",
                                        "example": "150000000"
                                    },
                                    "race": {
                                        "type": "string",
                                        "example": "Saiyan"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "example": "Male"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Goku después de entrenar..."
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "https://ejemplo.com/goku_nuevo.webp"
                                    },
                                    "affiliation": {
                                        "type": "string",
                                        "example": "Z Fighter"
                                    },
                                    "originPlanet": {
                                        "description": "Opcional. Si lo envías, debes mandar sus datos completos.",
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Tierra"
                                            },
                                            "isDestroyed": {
                                                "type": "boolean",
                                                "example": false
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Hogar adoptivo de Goku"
                                            },
                                            "image": {
                                                "type": "string",
                                                "example": "https://ejemplo.com/tierra.webp"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "transformations": {
                                        "description": "Opcional. Si lo envías, borrará las anteriores y guardará estas nuevas.",
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Ultra Ego"
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "https://ejemplo.com/ssj2.webp"
                                                },
                                                "ki": {
                                                    "type": "string",
                                                    "example": "150000000"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Personaje actualizado con éxito"
                    },
                    "404": {
                        "description": "Personaje no encontrado"
                    },
                    "422": {
                        "description": "Error de validación (ej. El planeta no es válido)"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Personajes"
                ],
                "summary": "Eliminar un personaje",
                "description": "Realiza un borrado lógico (Soft Delete) del personaje indicado por su ID.",
                "operationId": "f49405bd769a339c74f76d42feb0d79e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "El ID numérico del personaje que vas a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Personaje eliminado con éxito"
                    },
                    "404": {
                        "description": "Personaje no encontrado (o ya estaba eliminado)"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Personajes",
            "description": "Personajes"
        }
    ]
}